Fix bug in draw_samples_CP.m
Due to how the cell term_2 is defined, all entries of term_2 are products that incorporate the Phi matrix. Therefore, it is incorrect to incorporate the Phi matrix again the way I did on the old line 65 when I computed common_terms for the sampling of the first sub-index. The calculations for subsequent sub-indices appear to be correct.